backend/requirements-api.txt
# Lightweight requirements for API-based embeddings (no PyTorch)

# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# HTTP requests for API calls
requests==2.31.0

# Vector stores & search
faiss-cpu==1.8.0
chromadb==0.4.22
rank-bm25==0.2.2

# Language processing
nltk==3.8.1
stanza==1.7.0
indic-transliteration==2.3.44

# Topic modeling (optional, can be removed if not needed)
bertopic==0.16.0
umap-learn==0.5.5
hdbscan==0.8.33
scikit-learn==1.3.2

# LLM & RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.45
langchain-groq==0.1.3
openai==1.6.1

# Translation
langdetect==1.0.9

# Evaluation
evaluate==0.4.1
rouge-score==0.1.2

# OCR (optional)
easyocr==1.7.0

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Utilities
tqdm==4.66.1
loguru==0.7.2

# LiveKit Voice Agent
livekit-agents
livekit-plugins-openai
livekit-plugins-cartesia
livekit-plugins-silero
